Working days

+++ libft/stdlib/ft_atoi.c
@@ -6,15 +6,15 @@
 /*   By: welepy <welepy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/17 16:00:12 by marcsilv          #+#    #+#             */
-/*   Updated: 2024/11/11 18:24:21 by welepy           ###   ########.fr       */
+/*   Updated: 2025/03/10 17:48:54 by welepy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-int	ft_atoi(const char *str)
+long long int	ft_atoi(const char *str)
 {
-	int	num;
+	long long	num;
 	int	isneg;
 	int	i;
